Assert that user owns comment before deleting it

+++ api/src/main/comment/router.py
@@ -5,7 +5,8 @@ from starlette.status import HTTP_200_OK, HTTP_201_CREATED, HTTP_204_NO_CONTENT
 from src.main.comment import crud
 from src.main.comment.schema import CommentCreate
 from src.main.comment.settings import settings
-from src.main.comment.util import get_username_from_access_token, assert_post_exists
+from src.main.comment.util import get_username_from_access_token, assert_post_exists, \
+    assert_user_is_owner_of_comment
 from src.main.shared.database.main import get_db
 
 router = APIRouter(prefix=settings.SERVICE_PREFIX)
@@ -28,6 +29,9 @@ def create_comment(request: Request, comment: CommentCreate, db: Session = Depen
 
 
 @router.delete("/{comment_id}", status_code=HTTP_204_NO_CONTENT)
-def delete_comment(comment_id: int, db: Session = Depends(get_db)):
-    # TODO: assert user is owner of comment
-    return crud.delete_comment(db=db, comment_id=comment_id)
+def delete_comment(request: Request, comment_id: int, db: Session = Depends(get_db)):
+    assert_user_is_owner_of_comment(db=db, request=request, comment_id=comment_id)
+
+    crud.delete_comment(db=db, comment_id=comment_id)
+
+    return
